GC: Added tests where unreferenced node adds reference to other unreferenced nodes (#13296)
Currently, when an unreferenced node adds reference to other
unreferenced nodes, their unreferenced timestamps are not updated.
However, this is incorrect because any client can retrieve the objects
for the nodes whose references are added and then they have until
session expiry to reference it back. So, the unreferenced timestamp
needs to be updated to give enough time for these sessions to expire and
we can safely delete these objects.

This PR adds tests for these scenarios and validates that they are
failing currently. Once the bug is fixed, the tests will be updated so
that they pass.

Bug for the above scenario -
[AB#2801](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2801).