Add logging for when the connection mode granted is different than th… (#13302)
While investigating [Task
2342](https://dev.azure.com/fluidframework/internal/_workitems/edit/2342)
we've realized that we request a connection mode to the server and if it
does not grant it, we do not log at all. That makes it much harder to
understand why the summarizer was being connected in read mode instead
of write. Adding a telemetry event to catch these cases.