Support for views in IsoBuffer. (#13252)
This adds support for ArrayBuffer views to the IsoBuffer. 

It removes an assertion that was previously triggered, if a view was
used and adds a test to demonstrate that views work as expected.

This fix is needed to prevent an exception in PropertyDDS, which uses
the slice command to chunk the summary.