GC tests where summarizer loads from an older snapshot (#13328)
Added tests which repros a bug where GC data is not correctly update on
refreshLatestSummaryFromSnapshot:

refreshLatestSummaryFromSnapshot downloads a specific snapshot from the
server and various latest summary state is updated from this snapshot.
However, the GC data in summarizer nodes is not updated from this
snapshot which can lead to incorrect behavior. For example, consider the
following scenarios:
A data store (say ds1) is unreferenced in summary1 at seq#100.
An op references ds1 and it is referenced in summary2 at seq#101.
A new summarizer loads from summary1 (seq#100). While catching up to the
latest, gets ack for summary2 and calls refreshLatestSummary. The
referenceSequenceNumber for ds1 is updated to seq#101.
In the next summary (summary3), if ds1 did not change, it will not run
GC again since its referenceSequenceNumber is the same since last
summary (summary2).
GC will use data from previous summary which is the data from summary1.
This is because GC did not update data from summary2. In this summary,
ds1 will be unreferenced incorrectly.

Basically, any data stores that did not change between
refreshLatestSummaryFromSnapshot and the next summary may have incorrect
GC data.

Demonstrates the bug in
https://dev.azure.com/fluidframework/internal/_workitems/edit/2821.