GC: Fix for unreferenced timestamp not always reset when reference is added between summaries (#13316)
## Bug
When a reference is added to unreferenced objects between two summaries,
their unreferenced timestamp is reset only if the reference is added
from another referenced object. See
https://github.com/microsoft/FluidFramework/blob/3cd0cd0618b2963b2e24261da9a4ae6794c90d4e/packages/runtime/container-runtime/src/garbageCollection.ts#L1309.
However, the unreferenced timestamp of such objects (say objA) should be
reset irrespective of whether the object that adds the reference (say
objB) is itself referenced or not. That's because objB can be in memory
of certain clients even if it is unreferenced and such clients could
then access objA and store an in-memory reference to it. The session for
these clients can last upto the session expiry timeout and so, their
unreference timestamp needs to be reset.

## Fix
When updating the unreferenced timestamp of references between
summaries, do it for all all references irrespective of whether the node
that added the reference was referenced or not.


[AB#2801](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2801)