Gitrest: initial write speedup (#13249)
## Description

When using a remote filesystem, the overhead of filesystem operations
makes shredded summaries extremely slow to read/write. We already work
around this for Read by computing and storing a latest "Full Summary"
single file that is used on read for most "Latest" summary reads.

The only Summary Write operation in the critical path is when creating
the initial summary for a new document. This PR attempts to reduce the
overhead of a shredded summary read/write by "compressing" the git tree
for the initial summary into a single file (using Isomorphic git with
`memfs`).

## Breaking Changes

N/A, only impacts new documents

## Reviewer Guidance

I validated the following scenarios to ensure functionality with each
`enableLowIoWrite` setting ("initial", true, false) as well as
forward/backwards compatibility using local R11s and my fluid app.
- Write service summary from initial summary
- Write client summary from initial summary from first client
- Write client summary from initial summary from 2nd client (no service
summary in the middle)
- Write service summary after client summary(ies)
- Write client summary after service summary(ies)
- Load from service summary
- Load from initial summary
- Load from client summary