Corrupted merge tree (#13301)
Fix for the merge tree bug where a node with no children and no parent
remained in the tree and was the reference for `EndOfTreeSegment`. 2
main things contributed to this problem:
1. The `EndOfTreeSegment` had a different view of the root than the rest
of the tree. This was due to the root of the tree being updated at some
point, but since the `EndOfTreeSegment` only computed the root once at
the time of construction, it saw the root having a different number of
children than the actual root had.
2. The `packParent` method attempted to pack nodes when it should not
have (when there were no children to pack). In this case, execution
should jump straight to the recursive call on the parent and not create
a parent node for a node with no children.


[AB#1919](https://dev.azure.com/fluidframework/internal/_workitems/edit/1919)