Add Regression Test for Reported Undo/Redo Regression (#13332)
There was a reported regression in the existing/old undo redo stack
around deleted segments, which are known to have problems. Added a
regression to the new revertible infa which should resolve the problems
around deleted segments. The issue does involve marker segments which
should not matter, as both the old and new infra is agnostic to marker
types. This was confirmed by the regression tests


[AB#2720](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2720)

Co-authored-by: Tony Murphy <anthonm@microsoft.com>