Add SharedCounter editor control to client-debug-visualizer package (#13350)
#### Description 

https://dev.azure.com/fluidframework/internal/_workitems/edit/2656/

#### Objective 

<img width="911" alt="image"
src="https://user-images.githubusercontent.com/111468570/207467001-72a3fbcb-4138-4f4c-9209-86ebdd5fdd0e.png">

This PR aims to add functionality so that the users can freely
manipulate the `sharedCounter` values by incrementing, decrementing or
by passing an arbitrary integer value in the input field that would
function as a delta value.


#### Notes 
- There should be some CSS works to make the visualization neater after
the logic for adding `sharedCounter` editor is approved.
- The current method of adjusting `sharedCounter.value` is by repeating
`sharedCounter.increment` for `num - value` number of times. We might
have to think about coming up with a more efficient solution or changing
the design of `sharedCounter` to allow applying an arbitrary integer
value in the future.