Support content attribution on merge-tree (#12634)
## Description

Adds functionality to merge-tree which supports persisting the sequence
numbers associated with a given segment. At the time of insertion, a
segment is associated only with the `sequenceNumber` at which it is
inserted, but over time as segments are merged/split, a segment may
comprise a number of different sequence numbers at its different
offsets.

This functionality is gated by a merge-tree option which can be turned
on through data store runtime options.

This PR also adds a test suite to SharedString which helps assess
attribution's impact on document size when using various serialization
techniques.