Explore creating a signal to containers that an external data source has changed (#13063)
**EDIT: This is a throw away PR and is not intended to be merged into
main. We are simply using this branch as a playground for hacking around
in for now.**

This PR explores creating a signal that will be emitted from the server
to the container in the client that the source of truth for the data
that the client has changed and that the container in the client should
do something with that information.

In order to try it out yourself, please do the following:
1. Start a codespace of branch (or pull it down to you linux box)
2. Navigate to protocol-definitions, install packages and build them: 
`$ cd common/lib`
`$ npm i`
`$ npm run build`
3. Go back to root, and link to the protocol-definitions:
`$ cd ../..`
`$ npm link common/lib/protocol-definitions`
4. Install the rest of the node packages and build them at root:
`$ npm i`
`$ npm run build`
5. Navigate to routerlicious, link to protocol-definition, then install
and build the libraries there:
`$ cd server/routerlicious`
`$ npm link ../../common/lib/protocol-definitions`
`$ npm i`
`$ npm run build`
4. Navigate to `server/routerlicious/packages/lambdas` and link from
there because the package name and the directory name are different:
`$ cd packages/lambdas`
`$ npm run build`
`$ npm link`
5. Navigate to the tinylicious directory from routerlicious, install and
link all pacakges and then build:
`$ cd .../../tinylicious`
`$ npm i` 
`$ npm link ../../common/lib/protocol-definitions/`
`$ npm link @fluidframework/server-lambdas` (this completes the link
started in the last part of step 6.)
`$ npm run build`
6. Run tinylicious: `$ npm run start`
7. Navigate to the example directory and run it:
`$ cd ../../examples/hosts/app-integration/external-data`
`$ npm run start`

Now you should be able to go to a browser at `localhost:8080` and turn
on the developer tools, and click on the Network tab and Websockets
filter to see the WS. If it is empty, just refresh. You should see the
following signals exchange taking place in the Networks tab then, (note
the greyed externalDataChange signal):


![image](https://user-images.githubusercontent.com/6777404/204363677-54dc3ff1-ea4a-49fd-8ec6-7f0fc449d6cd.png)