Fix rollback farm with empty property comparison (#13401)
## Description

The rollback farm for mergeTree fails with annotateRange because setting
and rolling back properties results in empty propertySets which do not
compare as equal to undefined propertySets. This change adds another
check in the test code to compare {} and undefined property sets as
equal.

I had previously tried to address this in #11264 by changing the
matchProperties function to treat undefined and {} as equivalent, but
that broke snapshot tests in a way I was uncertain was safe, so I
reverted. Another possibility is to delete properties when they become
empty on the segment (#13400), also a breaking change.