Removed unrefTimestamp from IGarbageCollectionDetailsBase (#13382)
`unrefTimestamp` is not needed in `IGarbageCollectionDetailsBase` since
data stores do not write GC data to summary. However, older documents
may still have the GC data in this format in data store's summary tree.
So, moved the existing `IGarbageCollectionDetailsBase` interface to a
deprecated legacy interface that is only used when reading older
summaries.