When enabled, chunk compressed batches larger than 1MB (#13329)
Part of ADO:2089.

This adds the functionality of accepting batches which exceed the 1MB
limit after compression, making it possible to submit payloads of any
size. It is based on some of the previous chunking work which has been
decommissioned a while back as it became incompatible with the default
flush mode.

What currently happens is that if a batch is larger than the minimum
compression threshold and gets compressed, yet its post-compression size
is still larger than the maximum payload limit (1MB), the container will
close with an error, as the payload is going to be rejected by the
socket anyway.

This change leverages the form of the compressed batch (which has one
large message at the beginning and empty messages in the tail in order
to reserve sequence numbers) by chunking the first op in small pieces
which can be sent as separate ops (technically small one op batches) and
then sending a batch with the last chunk as the first op along with the
empty messages.

The feature is disabled by default, and it requires both compression to
be enabled (via `compressionOptions` in `IContainerRuntimeOptions`)
along with chunking (via specifying a value for `chunkSizeInBytes` in
`IContainerRuntimeOptions`).

**The feature is experimental and changes the document schema**. Sending
chunked messages has been disabled for a while, but the runtime is still
able to process them. The chunked messages added with this change will
require a runtime which is able to process them. Therefore, it is
recommended that the feature is enabled only after the version of this
change reaches full saturation, to ensure that all clients are able to
interpret and process the new schema. The config enabling the feature
(`chunkSizeInBytes` in `IContainerRuntimeOptions`) is relative only to
producing the ops and not consuming the ops. The consumer is able to
interpret these ops regardless of configuration.

This is part 3 of 3 from
https://github.com/microsoft/FluidFramework/pull/13131