Add alternative event API and use it in SharedTreeCore (#13284)
This introduces a very small eventing library in @fluid-internal/tree
which has some advantages over the existing event pattern used by the
Fluid framework. It tries to remain in the spirit of the existing
pattern. Some differences are:

* It uses property maps (`{ eventName: (args) => void }`) rather than
function interfaces/intersections (`{(event: "eventName", listener:
(args) => void): void}`) to define the events and listeners, which is
more concise.
* The `emit` method is strongly typed.
* The `on` method is stronger-ly typed; supplying an event that is not
part of the event/listener declaration is an error.
* The `emit` method is not public on event emitter implementations, i.e.
clients that subscribe to events can't also fire events....
* ...unless: using the static creation function on the reference
implementation (`EventEmitter`) which allows classes to compose over an
event emitter rather than having to subclass it, which is sometimes
inconvenient or impractical.
* Events are deregistered by invoking a callback that is returned during
registration. This "React-style" pattern is more concise and often more
convenient than using a separate `off` method. This also brings the
required methods for an event emitter to just one (`on`), making it less
boilerplate to implement.
* It does not require as many complicated type transformations and type
boilerplate as the existing pattern.

The new event registry pattern (property maps) is convertible to the old
pattern (function interfaces) via the provided `TransformEvents` type,
so users can write their events using the new pattern and easily
transform them to existing Fluid event emitters that require the old
pattern.

There are many existing npm packages that provide some of the above features,
but they don't have everything (namely the type conversion to
TypedEventEmitter's event type format and the pattern for inheritance
vs. composition). It is not very much code to implement this, so
although we could certainly go with a third party option I do not think
it's a bad idea to roll it ourselves for now, especially given that it retains
the familiar methods in the API (`on` and `emit`) and therefore should
not increase concept count for new adopters if this code were to be
eventually shipped as a public-facing API. If we do decide to adopt a
third party library instead at a later date, it should be relatively easy
to migrate since the API is consistent.

This PR uses the new API for `SharedTreeCore` and its `Indexes`. Indexes
do not currently deregister their events since their lifetime matches
that of the shared tree.