Gate storage of attribution data behind a config flag (#13418)
## Description

Introduces the `Fluid.Attribution.EnableOnNewFile` configuration, which
defaults to false. This config changes storage behavior of attribution
data on new files. Behavior for existing files in unchanged (if a
container runtime with mixed in attributor is used, attribution will be
enabled if and only if it was enabled when the file was created).

The main motivation for this change is it enables a plausible dark
rollout path for consumers, wherein they can safely start shipping a
version of the app which uses a container runtime with mixed in
attribution functionality, and only when they're ready turn on this
feature flag (and can do so incrementally).

This may not be the ultimate mechanism we suggest for rollout, we may
want something more resilient. Even in that event, this config flag
could be useful for some development flows, since merge-tree attribution
isn't currently pay-to-play (so runtime developers could enable it on
their boxes and experiment with some data they get).

## Reviewer Guidance

This is the first config flag I've introduced. Please pay closer
attention to non-idiomatic patterns or missing things in the changes
(ex: is there a place we're meant to document config types?).


[AB#2922](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2922)