R11s Driver: Add 5 day expiration to the in-memory cache for compatibility with GC (#12557)
Fixes
[AB#2238](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2238)

The Garbage Collection feature requires a limit on the age of the
snapshot used for loading in conjunction with a limit on the length of a
given session. These limits provide a cap on how long we need to wait
from when an object becomes unreferenced to when we know no client could
possibly still have a reference to it in memory.

The limit on snapshot cache age is reflected in the policy
`IDocumentStorageServicePolicies.maximumCacheDurationMs`. Unfortunately,
the current structure of how the driver and runtime are instantiated for
new documents precludes the reliable transmission of this policy from
the driver where it's implemented to the runtime where it's needed to
compute the Sweep Timeout.

We have a long-term plan to move ownership/transmission of this policy,
but in the meantime, we need to ensure all drivers adhere to the same
policy across the board. As-is, if an experience built on the R11s
driver attempted to enable GC, then in the case of a very stale cache,
document corruption could result.