Fixed bug where GC data is not updated in summarizer node on refreshLatestSummaryFromSnapshot (#13361)
## Bug
`refreshLatestSummaryFromSnapshot` downloads a specific snapshot from
the server and the latest summary state is updated from this snapshot.
However, the GC data in summarizer nodes is not updated from this
snapshot which can lead to incorrect behavior. Any data stores that did
not change between refreshLatestSummaryFromSnapshot and the next summary
may have incorrect GC data - GC will use the GC data from the previous
GC run which is outdated since it did not update from the latest
snapshot.

## Fix
Update the GC data from the downloaded snapshot when
`refreshLatestSummaryFromSnapshot` is called. The GC data is present at
the root of the snapshot tree under "gc" sub-tree. Here is how it is
used to update all summarizer node's GC data:
1. The root summarizer node gets the GC data from the root snapshot. It
runs GC on the data to generate used routes for the container.
2. The summarizer node stores the GC data and used routes as the ones
from latest summary. It then adds the GC data and used routes of its
children to their snapshot trees.
3. The child nodes receive the GC data and used routes and repeats step
2 above.

Note that the main and interesting logic is in
`summarizerNodeWithGC.ts`. The rest is mostly refactoring that is needed
for this change.


[AB#2821](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2821)