Changed gcUnknownOutboundReferences to error and added test to validate GC's  refresh from snapshot is correct (#13426)
Updated "gcUnknownOutboundReferences" to be logged as an error - It
should have been an error to begin with since its not expected.
Added couple of tests that show the bug where GC state is not updated
correctly in GarbageCollector on refreshing state from a snapshot.
Basically, the gc data from last run and tombstone state should also be
updated when refereshing state from a snapshot. Otherwise, there are
couple of things that can go wrong:
- The unreferenced timestamp of nodes is not updated in certain
scenarios when it goes from unreferenced -> referenced -> unreferenced
between summaries. Added a test for this scenario.
- "gcUnknownOutboundReferences" error can be logged in certain
scenarios. Added a test for this scenario.

These tests repro the issues in bugs
[AB#2940](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2940)
and
[AB#2945](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2945)