Resolve incorrect partial lengths state that prevented insert operation (#13439)
This PR addresses the insert bug that appeared in the rollbackFarm
tests. Insert ops that were created to be rolled back were being marked
as removed without ever being acked. markRangeRemoved updates the
partialLengths by subtracting the length of the node that was inserted,
but since the length was never added to the tree in the first place,
this resulted in negative node lengths.


[AB#1809](https://dev.azure.com/fluidframework/internal/_workitems/edit/1809)