Added a test to demonstrate a bug where de-duping a tombstoned blob makes it unusable (#13447)
Added a test that demonstrates the following bug:
After an attachment blob with id say `/_blobs/uniqueId1` has been marked
inactive / sweep ready / tombstone, if someone uploads a blob with the
same content, the server will give back the same id `/_blobs/uniqueId1`.
Now, adding this blob's handle or getting it will result in inactive /
sweep ready / tombstone error today.
However, when a blob is uploaded, its lifetime should be the same
whether it is de-duped or not. So, in the above scenario, there should
be no inactive / sweep ready / tombstone error.