Fixed bug where GC state in GarbageCollector is not correctly updated on refresh from snapshot (#13427)
`refreshLatestSummaryFromSnapshot` downloads a specific snapshot from
the server and the latest summary state is updated from this snapshot.
However, the GC state in GarbageCollector is not updated from this
snapshot which can lead to incorrect behavior. Couple of examples of
things that can be incorrect:
- The unreferenced timestamp of nodes is not updated in certain
scenarios when it goes from unreferenced -> referenced -> unreferenced
between summaries.
- "gcUnknownOutboundReferences" error can be logged in certain
scenarios.

When refreshing from a snapshot, the GC state and tombtsone state from
the last GC run is also updated from the downloaded snapshot. The state
and timers of currently unreferenced nodes is updated as per the GC data
from the snapshot. The tombstone state is also updated as per the
snapshot.