GC: Added test where a DDS's summarizer node is created after refreshLatestFromSnapshot completes (#13452)
Added a test that repros the following bug:
When a DDS's summarizer node is created after the container refreshes
its state from a snapshot, the DDS's GC data may be incorrect. Here is a
scenario where this can happen:
1. Container loads from base snapshot - snapshot1. The GC data for
dataStore1 is set to dsGCData1. It has a map DDS but the data store
hasn't been realized yet so the map's summarizer node is not created.
The map DDS's GC data at this point is mapGCData1.
2. Container gets an ack for snapshot2 and refreshes its state from this
snapshot. The GC data for dataStore1 is updated to dsGCData2. The map
DDS's GC data at this point is mapGCData2.
3. dataStore1 is now realized and the summarizer node for map DDS is
created. Currently, the map's GC data is initialized from the base
snapshot's GC data, i.e., mapGCData1.
4. The map DDS does not change and GC runs. The GC data returned in
mapGCData1 which is incorrect. In step 3, the map's GC data should have
been updated to mapGCData2.

Repros this bug -
[AB#2962](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2962)