Added tests for attachment blobs uploaded when container is disconnected (#13463)
Added tests for scenarios where attachment blobs are uploaded when
container is disconnected. When such a container is connected, the blobs
are uploaded to the server and a local to storage id mapping is created.
The code path is different from when blobs are uploaded in detached
container and the container is attached later. We need tests that cover
both these scenarios.
These tests also ensure that "gcUnknownOutboundReferences" errors are
not thrown for the local to storage id mapping.