Log when mismatched `referenceSequenceNumber` messages are detected by the batch manager - main (#13397)
For compression to work as expected and maintain batch correctness, it
is required that all messages within the same batch have the same
reference sequence number.

The feature was partially ported from `next` (without the API on the
runtime layer to block op creation from the event handler) and it is
disabled by default (see
https://github.com/microsoft/FluidFramework/pull/12326 and the
description for details). **Compression would enable it.**

By default, the feature is disabled. However, we'd like to know if and
when it happens, so this change will log at most 5 instances per session
along with the stacktrace in order to help us better understand the
customer scenarios and find a workaround.

Also fixing the issue with the AgentScheduler performing op reentry
(also ported from `next`).

Regardless of the runtime options (including the compression options),
when the feature is explicitly enabled, if the feature gate
`Fluid.ContainerRuntime.DisableOpReentryCheck` is set to `true`, the
feature will end up disabled.