Tree: Nest Core libs into core directory (#13487)
## Description

This moves the `core` libraries (which are aggregated into `core`, see
`src/core` and the tree package readme) into the core directory. This
makes the structure of the dependencies more clear and more enforceable.

This change was originally left out when introducing the core package to
reduce code churn, but this seemed like a good time to do it.

Also adds a missing fence file to id-compressor.

Also forest (and other indexes) no longer need to depend on the
abstractions in "shared-tree-core" (and haven't for a while) so
"shared-tree-core" was not moved into core and was separated from the
rest of the core libraries.