Remove partial lengths debug-only code from bundle (#13432)
The pattern used previously causes debug-only code to be included in the
bundle. We don't export `PartialSequenceLengths` from the package, so
there's no way consumers can (or should) turn this verification on. This
adjusts the options object to take in a verify callback instead, which
enables tree-shaking of those functions.