Fixed bug where DDS's summarizer node's GC data is not correctly updated on creation after refreshing from snapshot (#13455)
When a DDS's summarizer node is created after the container refreshes
its state from a snapshot, the DDS's GC data may be incorrect. This can
happen if the DDS did not change since refreshing from snapshot.
Currently, the DDS's GC data will be updated from the base snapshot the
container loaded from and if that's different from the GC data in
refreshed snapshot, the DDS will have incorrect GC data.

The core fix is that when a child summarizer node is created, the parent
node uses its current GC state to generate the base GC details for the
child node. This way, if the parent's GC state was updated because of
refresh from snapshot, the child will get the updated GC data.
Currently, the base GC data of a node is provided by the context that
creates it. With this change, the parent node will ignore this GC data
and instead update the child from its own GC data. Note that the GC data
still needs to be passed due to back-compat. This will be removed later.
Couple other key changes:
- ContainerRuntime passes in the base GC details to its summarizer node
during creation. This will then be passed on to the data stores's
summarizer node during their creation. The data store will pass it on to
the DDS's summarizer node during creation and so on.
- ContainerRuntime calls summarizer node's getGCData instead of directly
getting the GC data from DataStores. This will ensure that the container
runtime's summarizer node always has up-to-date GC data. This did not
happen before because its GC data was never used.
- Fixed a bug in `unpackChildNodesGCDetails` API which was changing the
passed GC data be removing node with id "/".

Fixes
[AB#2962](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2962)
Also, fixes
[AB#2917](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2917)
partly