feat: Add initial Fluid Client Debugger prototype libraries (#13425)
### Libraries

- @fluidframework/client-debugger
- Contains library for initializing / closing debuggers associated with
Container instances. The majority of this library's exports are
`@internal` intended to be used by related libraries (this policy can
always be changed in the future if we wish to allow external users to
write their own tooling against the types here).
- @fluidframework/client-debugger-view
- A visualizer library that displays basic stats about any debugger
instances registered with `@fluidframework/client-debugger`.
- @fluidframework/client-debugger-chrome-extension
- A thin wrapper around `@fluidframework/client-debugger-view`, which
generates an installable Chrome extension.

NOTE: All of the above libraries are still in early development. They
are all marked as private, and will not be published until they have
been deemed feature complete.

Co-authored-by: Tong Chen <114451900+chentong7@users.noreply.github.com>
Co-authored-by: jikim-msft <111468570+jikim-msft@users.noreply.github.com>