stress: re-enable offline (#13513)
Re-enable offline in the stress tests, which were disabled due to errors
suspected to be caused by offline. I have not been able to reproduce the
errors in test branch runs, so I would like to re-enable it and monitor.
It's possible that the errors were actually caused by offline, but only
occur when the servers are under load.