Upgrade the GC version as we found some bugs which means current GC data could be incorrect (#13532)
Recently, the following issues with how GC data is generated and added
to the summary were fixed: #13361, #13427 and #13455. This means that
the GC data in existing documents could be incorrect. Upgrading the GC
version will regenerate the GC data in these documents.
The upgrade is behind a feature flag and can be rolled out in stages.