Update client lock file for 2.0.0-internal.2.2.1 release (#13565)
## Description

This updates the lerna-package-lock.json file for the client release
group.

This was generated by running `npm i`.

Relevant npm bugs:
- newly release packages cause changes in the package lock due to
handling of aliased packages. This includes those changes.
- package hashes pick up which ever format was present in the user's
node_modules folder. Mine has sha512 for a few packages which has sha1
versions currently committed. This is an improvement in security and is
the default, so I'm considering this a fix of a previous regression. See
https://github.com/microsoft/FluidFramework/pull/13425/files#diff-af7f3050477fb02cc2cd3e3825846a1985372da54ec750a8cd7aea241f15ce8fL45413

Co-authored-by: Craig Macomber <Craig.Macomber@microsoft.com>