Refactor compressor utilities (#13564)
## Description

As part of moving the IdCompressor into the runtime, we'll need to drag
and drop it into the container-runtime. Right now, the compressor is
somewhat coupled to SharedTree's util module. This moves the utilities
it needs inside the id-compressor so that the utilities can move with
the compressor. I've also removed unused utilities from SharedTree.