Attachment blobs - Always return local ID and send blob attach op for all blob uploads including de-dupe scenarios (#13486)
## Problem
There are 2 main reasons for adding this change and both are because of
GC:
1. The main reason this change is added because of the way blob de-dupe
interacts with GC. If a blob that has been unreferenced for a long time
(almost ready for deletion) is de-duped, the server returns the same
storage ID for it as the unreferenced blob. GC doesn't know about that
the blob was de-duped and will proceed to delete the blob shortly.
However, the client that got the de-duped storage ID can still refer to
the blob leading to data loss.
The behavior in the case of blob de-dup should be the same as a brand
new blob upload since from the users perspective thqat is the case. So,
when the blob is de-duped, the time to delete it should be reset giving
the user the same amount of time to re-use the blob as with a brand new
blob upload.

2. The other reason is that GC validates all references from a node to
another are explicitly notified. In case of blobs, a mapping from local
ID to storage ID is essentially a reference and GC should be notified of
it.

## Solution
The high level solution is that when a blob is de-duped, notify GC so
that it resets the unreferenced timestamp. Here is how this PR achieves
that:
- It always returns local IDs when a blob creation is completed.
Currently, local IDs are only returned in detached / disconnected mode.
This makes the creation flow consistent and easy to reason about - When
a blob creation is complete, a local ID to storage ID mapping is created
and the local ID is returned to users.
- Whenever a local ID to storage ID mapping is created, notify GC that a
reference is added. This will fix problem 2 above and since a mapping is
created in blob de-dupe scenario as well, it will fix problem 1 as well.

## Reviewer Guidance
Couple of key changes to note:
- When container is online and a blob upload is completed, before the
local ID is returned, a blob attach op is sent so that all clients can
create a local ID to storage ID mapping on receiving it. This will
ensure that any op containing this local ID is sequenced after the blob
attach op and clients can retrieve the blob successfully.
- When a blob is de-duped and the original blob is already ack'd, the
blob is returned via the local ID as soon as the upload completes and
the blob attach op is sent. Since the blob is already present, we don't
need to wait for the op to be ack'd as the server will keep it alive.
- When one or more blobs are de-duped and the original blob is not yet
ack'd, these ops are added to a local queue (opsInFlight). As soon as
one of the ops is ack'd, all the blobs are resolved and returned since
this op will keep all of them alive as they have the same storage ID.



[AB#2745](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2745)