Fix for GC tests that load from old summaries not working for ODSP (#13583)
## Bug
The GC tests in `gcLoadingFromOlderSummaries.spec.ts` validate that when
summarizer loads from an older summary and refreshing is state from a
later summary ack, things work as expected. In ODSP, when a summary is
uploaded the previous summary may get deleted as they only keep around 2
summaries for a document.
The tests generated 2 summaries (s1 and s2) and then loaded a summarizer
container from the s1. ODSP deletes s1 when s2 is submitted so the
container loads fails.

## Fix
Load the container from a summary before submitting another summary. In
the above scenario, load the summarizer container from s1 before s2 is
submitted. However, this results in another problem where the new
summarizer container is not chosen as the summarizer due to another bug.
To work around this issue, after the first summarizer is closed,
disconnect and reconnect the new summarizer. This adds the summarizer to
the quorum which results in summarizer election where its elected.


[AB#3103](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3103)