Remove compression from the message metadata (#13497)
## Description

ADO:2942

It is safe now to remove the `compression` flag from the message
metadata and leave only the top level `compression` message field.

## Breaking Changes

This is technically a breaking change, as new runtimes will not be able
to process compressed payloads from the older runtimes and we could deal
with this. However, this is not a shipped feature and also an
experimental feature, therefore it is safe to abruptly end support for
such a short-lived message protocol.